Add license, inline detector into locator
