+++ app/domain/import/person_duplicate/attributes.rb
@@ -1,3 +1,9 @@
+# frozen_string_literal: true
+
+#  Copyright (c) 2024, Schweizer Alpen-Club. This file is part of hitobito and licensed under the
+#  Affero General Public License version 3 or later. See the COPYING file at the top-level directory
+#  or at https://github.com/hitobito/hitobito.
+
 module Import::PersonDuplicate::Attributes
   extend ActiveSupport::Concern
   DUPLICATE_ATTRIBUTES = [
